\name{kandi_stuff-package}
\alias{kandi_stuff-package}
\alias{kandi_stuff}
\docType{package}
\title{
\packageTitle{kandi_stuff}
}
\description{
\packageDescription{kandi_stuff}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kandi_stuff}
\packageIndices{kandi_stuff}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{kandi_stuff}

Maintainer: \packageMaintainer{kandi_stuff}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
